Convert main entry to TypeScript
